+++ java/Crawler/src/main/java/crawler/ctrl/ContentManager.java
@@ -1,4 +1,4 @@
-package ctrl;
+package crawler.ctrl;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
