+++ java/Crawler/src/main/java/crawler/filesys/Downloader.java
@@ -1,8 +1,8 @@
-package folder;
+package crawler.filesys;
 
-import ctrl.ContentManager;
-import ctrl.MessageSender;
-import datalake.DataLake;
+import crawler.ctrl.ContentManager;
+import crawler.ctrl.MessageSender;
+import fsystem.DataLake;
 import org.jsoup.Connection;
 import org.jsoup.Jsoup;
 import java.io.IOException;
@@ -16,24 +16,23 @@ public class Downloader {
 
     public Downloader(DataLake dl){
         dataLake = dl;
-        System.setProperty("log4j.configurationFile", "log4j2.properties");
     }
     public void run() {
         try {
-                Random random = new Random();
-                int randomNumber = random.nextInt(70000) + 1;
-                String numStr = String.valueOf(randomNumber);
-                String bookUrl = "https://www.gutenberg.org/cache/epub/" + numStr + "/pg" + numStr + ".txt";
+            Random random = new Random();
+            int randomNumber = random.nextInt(70000) + 1;
+            String numStr = String.valueOf(randomNumber);
+            String bookUrl = "https://www.gutenberg.org/cache/epub/" + numStr + "/pg" + numStr + ".txt";
 
-                if (dataLake.isBookInDataLake(numStr)) {
-                    logger.info("Book is already downloaded.");
-                }else{
-                    if (downloadBook(bookUrl, numStr)) {
-                        logger.info("Download completed.");
-                    } else {
-                        logger.error("Book not found: " + bookUrl);
-                    }
+            if (dataLake.isBookInDataLake(numStr)) {
+                logger.info("Book is already downloaded.");
+            }else{
+                if (downloadBook(bookUrl, numStr)) {
+                    logger.info("Download completed.");
+                } else {
+                    logger.error("Book not found: " + bookUrl);
                 }
+            }
         } catch (Exception e) {
             e.printStackTrace();
         }
