+++ java/DataLake/src/main/java/cloud/CloudDatalake.java
@@ -1,7 +1,5 @@
 package cloud;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.auth.Credentials;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.cloud.storage.Blob;
