+++ java/DataLake/src/main/java/fsystem/DataLake.java
@@ -1,4 +1,4 @@
-package datalake;
+package fsystem;
 
 import java.io.BufferedWriter;
 import java.io.File;
