+++ java/DataLake/src/main/resources/credentials.json
@@ -1,0 +1,13 @@
+{
+  "type": "service_account",
+  "project_id": "search-engine-bd",
+  "private_key_id": "42126d0364d9a8ab7b21ee1b06336d47c70a7d35",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQC3WkvP1Ka/pX77\nlIIqVy6nOcGUDbLI8vxeVegepfixZ6xuI9GNxtwrHMniIaeMFjqC/dge/BtctpRi\nAwDyapUddLHyLWkxkhW+4LYdeEnIzG2iYheL+eNJWYj+zLDtAWGqYHyb5dQLGJdB\nPBlpX+V9LNOUntJxJ1SQVXLiQIl67ccUzAFMm8lgoWDNPgCI7q3g4QyGJKEzgARz\nlY1JWWRwjNZQWs1dt6XjsNs3dA/SoWz8S6z2+4opbcF44I/tTSi2uXPttEdHMWq7\n26WJZr7I9LyB7kX8cctynbfL1mjq8ZsY5dtlJ5dXJrvObvUDK3tkkpzQQPa4SoXU\nFz+sSXlxAgMBAAECggEAAaEaaRyCRlPzyMvJK9PmQNtrKFQGVZGOM07nXwVZzEbH\n8hscoLpGiNYKx7kFloG4j6MulVjna9eeAjB1ZNIUZOR9c9HaPFPoYE+6IpEn87I4\nW6/sNmTYM5XJeT0rqdGAMw3EMbT1QUt8xZpBfg2Zehcowqp2I4YGUvmFabQJ3Lcm\nGp5fMdTnFRIxfFEIibUjzzXYu5rXJ+bQyfmKnjBbAl+GLzQHzH7RNJupR5z6nkco\nPhRqh/+PpAkvxNA1GJ04XMTg/gx1jTEzW22OTnqZVt9jaHWOC8ls5JCMMFqAnin0\niwccJHu7V+Nds41NI6FstptPhhLTU4FWc/8DOA30+QKBgQDdy/njZ/BqSkMS11MK\naf0vUjKr+yS0oSkBxqKIC0TZQ/peRbyGw2AVjbhCEhMYBd0HRcUl4FJkBFZZly8h\ngf8vJluBsezt2blrGvug6ySvyh5g4hirrCMypykjbI9/47WNCpZnnT9f9VkSNp0S\nvBReRxF6NGitRCCjUtEKgvBMSQKBgQDToKMp8nkv9DwaMiHAxVsYh+iziO83I4YE\nQFZ9T6zTYmfqETo3NM2cyWtSo0Js48dpEg2YNvvy1hkVykvDckD2b8l5KUhgcwR6\nNut0M3tWtt21F0BdAcy7an/XnccrjSoNEwYFykUnVvEQb4rOzSJIIBORUuLTcjkt\nsO3B9lWz6QKBgGc0RaeyMqVSvtkihope/SeqyNdLui2um+Uup2xEShoKzpJVQr82\nZBxDYIyl2mqgN9PtfCn+snfFqORxO8YMtlAbOhMkVonqeKPE33dmUyDVHMOaET64\nj3w/Wh4YosEwlMb3OTsnaVjeUj7uWLJh/9xKl3BUuKx++ezlWU3Z1bS5AoGBALw6\nrcWakGWxs1+NZW4jW6IXCEffrP17EDFp1EPDATJG3axYRFJkGdmxYU7wPR3C/1Oa\nJhNUHIZiHY+/DeI6FkPbK/UjXefDhaD6hjQlsjdvkQUIOExBVIsB79Gk2UkIVvWo\nDPxLa3TRJLr8qxbDaBGEzkK82Vk+yFK2gtK5OoM5AoGBAL6WF/2EawCPQ3XK62i+\nUo15IcU5FWnT5YA1SIbn7SeqiyLZRhmq3NeExsjVG5t3CN9AVbWfHuB+FEpxKBPT\nhn3fIT0XgaajGbRCA1DrsxiVDdtv37euBGxTlIQPAGw8EUQr8kXjsanC085pydnu\nF3hSahE+Pv4lu/rZjgMklDj2\n-----END PRIVATE KEY-----\n",
+  "client_email": "carlosbigdata@search-engine-bd.iam.gserviceaccount.com",
+  "client_id": "107651316459116743786",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/carlosbigdata%40search-engine-bd.iam.gserviceaccount.com",
+  "universe_domain": "googleapis.com"
+}
